requirements.txt
fastapi
uvicorn[standard]
httpx
pandas
numpy
openai
python-dotenv

